requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.0
openpyxl==3.1.2
PyPDF2==3.0.1
pydantic==2.4.2 